% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey}
\alias{ex_survey}
\title{ex_survey: Mockup dataset of a survey.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 32 columns.
}
\usage{
ex_survey
}
\description{
A dataset containing fake respondents' answers to survey questions.
The first two, x_sex and x_human, are intended to be independent variables,
whereas the remaining are dependent. The underscore _ in variable names
separates item groups (prefix) from items (suffix) (i.e. a_1-a_9 => a + 1-9),
whereas ' - ' separates the same for labels.
The latter corresponds with the default in SurveyXact.
}
\keyword{datasets}
