% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsaver.R
\name{ggsaver}
\alias{ggsaver}
\title{Wrapper Function for \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
\usage{
ggsaver(plot, filename, ...)
}
\arguments{
\item{plot}{Plot}

\item{filename}{Note}
}
\description{
This only exists to make it easy to use it in \code{\link[=make_link]{make_link()}}
}
\examples{
library(ggplot2)
my_plot <- ggplot(data=mtcars, aes(x=hp, y=mpg)) + geom_point()
make_link(my_plot, folder=tempdir(), file_suffix = ".png", save_fn = ggsaver)
}
