% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_settings.R
\name{global_settings_set}
\alias{global_settings_set}
\title{Get Global Options for saros-functions}
\usage{
global_settings_set(new, fn_name = "makeme")
}
\arguments{
\item{new}{List of arguments (see \code{?make_link()}, \code{?makeme()}, \code{fig_height_h_barchart()})}

\item{fn_name}{String, one of \code{"make_link"}, \code{"fig_height_h_barchart"} and \code{"makeme"}.}
}
\value{
Invisibly returned list of old and new values.
}
\description{
Get Global Options for saros-functions
}
\examples{
global_settings_set(new=list(digits=2))
}
