% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeme.R
\name{make_content}
\alias{make_content}
\title{Method for Creating Saros Contents}
\usage{
make_content(type, ...)
}
\arguments{
\item{type}{\emph{Method name}

\verb{scalar<character>} with a class named by itself.

Optional string indicating the specific method. Occasionally
useful for error messages, etc.}

\item{...}{\emph{Dots}

Arguments provided by \code{makeme}}
}
\description{
Takes the same arguments as \code{makeme}, except
that dep and indep in make_content are character vectors,
for ease of user-customized function programming.
}
