% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link_global_settings.R
\name{make_link_global_settings_get}
\alias{make_link_global_settings_get}
\title{Get Global Options for \code{make_link()}}
\usage{
make_link_global_settings_get()
}
\value{
List with options in R
}
\description{
Get Global Options for \code{make_link()}
}
\examples{
make_link_global_settings_get()
}
