% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link_global_settings.R
\name{make_link_global_settings_reset}
\alias{make_link_global_settings_reset}
\title{Reset Global Options for \code{make_link()}}
\usage{
make_link_global_settings_reset()
}
\value{
Invisible returned list of old and new values.
}
\description{
Reset Global Options for \code{make_link()}
}
\examples{
make_link_global_settings_reset()
}
