% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link_global_settings.R
\name{make_link_global_settings_set}
\alias{make_link_global_settings_set}
\title{Set Global Options for \code{make_link()}}
\usage{
make_link_global_settings_set(new)
}
\arguments{
\item{new}{List of arguments (see \code{make_link()})}
}
\value{
Invisible returned list of old and new values.
}
\description{
Set Global Options for \code{make_link()}
}
\examples{
make_link_global_settings_set(new=list(digits=2))
}
