% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeme.R
\name{makeme}
\alias{makeme}
\title{Method for Creating Saros Contents}
\usage{
makeme(type, ...)
}
\arguments{
\item{type}{\emph{Method name}

\verb{scalar<character>} with a class named by itself.

Optional string indicating the specific method. Occasionally
useful for error messages, etc.}

\item{...}{\emph{Dots}

Arguments provided by \code{sarosmake}}
}
\description{
Takes the same arguments as \code{sarosmake}, except
that dep and indep in makeme are character vectors,
for ease of user-customized function programming.
}
