% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeme_global_settings.R
\name{makeme_global_settings_get}
\alias{makeme_global_settings_get}
\title{Get Global Options for \code{makeme()}}
\usage{
makeme_global_settings_get()
}
\value{
List with options in R
}
\description{
Get Global Options for \code{makeme()}
}
\examples{
makeme_global_settings_get()
}
