% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeme_global_settings.R
\name{makeme_global_settings_reset}
\alias{makeme_global_settings_reset}
\title{Reset Global Options for \code{makeme()}}
\usage{
makeme_global_settings_reset()
}
\value{
Invisible returned list of old and new values.
}
\description{
Reset Global Options for \code{makeme()}
}
\examples{
makeme_global_settings_reset()
}
