% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarosmake_global_settings.R
\name{makeme_global_settings_set}
\alias{makeme_global_settings_set}
\title{Set Global Options for \code{makeme()}}
\usage{
makeme_global_settings_set(new)
}
\arguments{
\item{new}{List of arguments (see \code{makeme()})}
}
\value{
Invisible returned list of old and new values.
}
\description{
Set Global Options for \code{makeme()}
}
\examples{
makeme_global_settings_set(new=list(digits=2))
}
