% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rng.R
\name{n_range}
\alias{n_range}
\title{Tidyselect-version of \code{\link[=n_rng]{n_rng()}}}
\usage{
n_range(
  data,
  dep,
  indep = NULL,
  glue_template_1 = "{n}",
  glue_template_2 = "[{n[1]}-{n[2]}]"
)
}
\arguments{
\item{data}{Dataset}

\item{dep, indep}{Character vector, names of (in)dependent variables}

\item{glue_template_1, glue_template_2}{String, for the case of a single
value (1) or a range with minimum-maximum of values (2).}
}
\value{
String.
}
\description{
Tidyselect-version of \code{\link[=n_rng]{n_rng()}}
}
\examples{
n_range(data=ex_survey, dep=b_1:b_3, indep=x1_sex)
}
