% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rng.R
\name{n_rng}
\alias{n_rng}
\title{Obtain range of N for a given data set and other settings.}
\usage{
n_rng(
  data,
  dep,
  indep = NULL,
  glue_template_1 = "{n}",
  glue_template_2 = "[{n[1]}-{n[2]}]"
)
}
\arguments{
\item{data}{Dataset}

\item{dep, indep}{Character vector, names of (in)dependent variables}

\item{glue_template_1, glue_template_2}{String, for the case of a single
value (1) or a range with minimum-maximum of values (2).}
}
\value{
Always a string.
}
\description{
Obtain range of N for a given data set and other settings.
}
\examples{
saros.contents:::n_rng(ex_survey, dep="b_1", indep = "x1_sex")
}
\keyword{internal}
