% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_sarosplot.R
\name{sarosmake_global_settings_reset}
\alias{sarosmake_global_settings_reset}
\title{Reset Global Options for \code{sarosmake()}}
\usage{
sarosmake_global_settings_reset()
}
\value{
Invisible returned list of old and new values.
}
\description{
Reset Global Options for \code{sarosmake()}
}
\examples{
sarosmake_global_settings_reset()
}
